refactor: integrate mkv-episode-matcher into engram codebase
- Remove dead code: engine.py (standalone orchestrator), config_manager.py
  (JSON config system), unused provider wrappers (Addic7edProvider,
  OpenSubtitlesProvider, CompositeSubtitleProvider), dead utility functions
- Eliminate duplicate Config class from core/models.py (engram uses AppConfig
  in SQLite via get_config_sync)
- Consolidate duplicate SubtitleReader, detect_file_encoding,
  read_file_with_fallback into single source in srt_utils.py
- Flatten directory structure: remove core/ nesting, rename files for clarity
  (core/matcher.py -> multi_segment_matcher.py, core/utils.py -> srt_utils.py,
  core/providers/asr.py -> asr_provider.py, etc.)
- Move investigation scripts to backend/scripts/matcher_investigation/
- Move test_episode_identification.py to tests/unit/
- Replace Rich console print() calls with loguru logger
- Fix import sorting and remove unused imports

22 files -> 14 files. ~1,800 net lines removed.
All 326 tests pass, lint clean.
