+++ backend/app/matcher/__init__.py
@@ -1,8 +1,4 @@
-"""MKV Episode Matcher - Audio fingerprint-based episode identification.
-
-This module is adapted from the standalone mkv-episode-matcher project
-for integration with Unified Media Automator.
-"""
+"""MKV Episode Matcher - Audio fingerprint-based episode identification."""
 
 from app.matcher.episode_identification import EpisodeMatcher
