--- backend/app/matcher/scripts/__init__.py
@@ -1,1 +1,0 @@
-# Investigation scripts package
